+++ src/OakLab.Blazor.Navigation.Tests/TestNavigationManager.cs
@@ -4,17 +4,15 @@ namespace OakLab.Blazor.Navigation.Tests;
 
 public class TestNavigationManager : NavigationManager
 {
-    private const string dummyUri = "https://www.google.com/";
+    public TestNavigationManager(string uri = "https://www.google.com/")
+    {
+        Initialize(uri, uri);
+    }
 
     public string? LastNavigatedUri { get; private set; }
 
     protected override void NavigateToCore(string uri, NavigationOptions options)
     {
         LastNavigatedUri = uri;
-    }
-
-    protected override void EnsureInitialized()
-    {
-        Initialize(dummyUri, dummyUri);
     }
 }